Commit des codes de la séance de travail ensemble du dimance 07-01-2018

+++ src/main/java/com/c353/bicomat/entities/Banque.java
@@ -17,7 +17,7 @@ public class Banque {
 	
 	@javax.persistence.Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
-	private Long id;
+	private Integer id;
 	
 	@Column
 	private String nom;
@@ -27,12 +27,12 @@ public class Banque {
 
 	
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 
@@ -59,7 +59,7 @@ public class Banque {
 	
 
 
-	public Banque(Long id, String nom, String adresse) {
+	public Banque(Integer id, String nom, String adresse) {
 		super();
 		this.id = id;
 		this.nom = nom;
